+++ src/Library.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberty\System;
+
+/**
+ * Class Library
+ */
+final class Library
+{
+    public const string VERSION = '0.0.1';
+}